rename message folder to lowercase
