+++ packages/app/src/lib/components/message/MessageReactions.svelte
