+++ packages/app/src/lib/components/message/MessageThreadBadge.svelte
